Fix reported issues (#2110)
* Fix: Check for empty string to avoid SIGSEGV

* Fix: avoid truncation and check for overflows on arithmetic operations

* Fix: warnings reported by ccc-analyzer

* avoid warning about unused variable, but actually used for assertions

* apply suggestions

+++ nasl/nasl_cert.c
@@ -161,7 +161,7 @@ nasl_cert_open (lex_ctxt *lexic)
   gpg_error_t err;
   tree_cell *retc;
   const char *data;
-  int datalen;
+  size_t datalen;
   ksba_reader_t reader;
   ksba_cert_t cert;
   object_desc_t obj;
